add socket command functionality basically refactored whole conferenceUtils

+++ src/storage.mjs
@@ -2,11 +2,10 @@ import fs from 'node:fs'
 import config from './config.mjs'
 
 export default function clearStorage() {
-    // ignore errors since it's most probably that they don't exist and that's fine
-    try {
+    if (fs.existsSync(config.STORAGE_CRYPTO_DIR)) {
         fs.rmSync(config.STORAGE_CRYPTO_DIR, { recursive: true, force: true })
-    } catch (e) {}
-    try {
+    }
+    if (fs.existsSync(config.STORAGE_BOT_FILE)) {
         fs.rmSync(config.STORAGE_BOT_FILE, { recursive: true, force: true })
-    } catch (e) {}
+    }
 }
